feat: seperated hooks for initial page
